Add support for a single domain URI scheme that works with the existing subdomain scheme used by tiddlyspace.
* A server request filter converts requests of the type:

http://host.com/space/some_space

to:

http://some_space.host.com/

* A server response filter converts uris in responses
to single domain form where necessary.

* The only modification to tiddlyspace code itself occurs
in the betaserialization module. the method _tiddler_as_div is
overridden so that environ['HTTP_HOST'] gets converted to single
domain form. Doing it there allows http_host to be set
before it is assigned to the server.host field of each
tiddler in the serializer.